toy/requirements.txt
absl-py==1.4.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
Farama-Notifications==0.0.4
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.3
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.34
glfw==2.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.12.2
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==8.3.1
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.2
jupyter-server-terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.3.0
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
mujoco==2.3.7
mujoco-py==2.1.2.14
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
# -e git+https://github.com/ChienFeng-hub/FlowQ.git@52ac430ecbd8fab01a61f1df6120e9f4c2bcb8a3#egg=normflows&subdirectory=normalizing-flows
notebook-shim==0.2.3
numpy==1.24.4
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proplot==0.9.7
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==5.3.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
# torch==1.13.0+cu117
# torchaudio==0.13.0+cu117
# torchvision==0.14.0+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3 #==2.0.4
wandb==0.15.9
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
werkzeug==2.3.7
widgetsnbextension==4.0.8
zipp==3.16.2
hydra-core
ray[tune]